fix(stock): correct StockLevelRepository query bugs
Three real bugs in findStockWithProductByWarehouse and
getGlobalStockByWarehouse:

1. The 'OR LOWER(p.sapCode) LIKE ...' branch sat OUTSIDE the
   parentheses of the search block, so a sapCode match
   returned the row regardless of warehouseId and
   currentStock > 0. Any product with matching sapCode in
   ANY warehouse leaked into the query results.
2. The countQuery did not include the sapCode branch, so
   totalElements in the Page was inconsistent with the data
   page. Now the count and the data query match.
3. getGlobalStockByWarehouse returned currentStock/pendingStock
   as Long via SUM(CASE WHEN ... THEN 1 ELSE 0 END), which
   counted the number of stock rows rather than summing the
   quantities. The projection interface declared them as
   Long instead of BigDecimal, and the WarehouseService
   mapper had to coerce. Replaced the COUNT-style sum with
   COALESCE(SUM(s.currentStock), 0) and the projection types
   are now BigDecimal end-to-end.

Part of Phase 2 (correctness bugs) of the audit.

+++ src/main/java/com/visco/backend/services/WarehouseService.java
@@ -554,14 +554,10 @@ public class WarehouseService {
           .warehouseId(p.getWarehouseId())
           .warehouseName(p.getWarehouseName())
           .totalStock(
-            p.getCurrentStock() != null
-              ? BigDecimal.valueOf(p.getCurrentStock())
-              : BigDecimal.ZERO
+            p.getCurrentStock() != null ? p.getCurrentStock() : BigDecimal.ZERO
           )
           .totalPendingStock(
-            p.getPendingStock() != null
-              ? BigDecimal.valueOf(p.getPendingStock())
-              : BigDecimal.ZERO
+            p.getPendingStock() != null ? p.getPendingStock() : BigDecimal.ZERO
           )
           .build()
       )
